Fix test

+++ e2etest/newe2e_task_runazcopy.go
@@ -125,6 +125,7 @@ type AzCopyEnvironment struct {
 	AzureClientId           *string `env:"AZURE_CLIENT_ID"`
 
 	InheritEnvironment bool
+	ManualLogin        bool
 }
 
 func (env *AzCopyEnvironment) generateAzcopyDir(a ScenarioAsserter) {
@@ -187,53 +188,55 @@ func (c *AzCopyCommand) applyTargetAuth(a Asserter, target ResourceManager) stri
 		// Only set it if it wasn't already configured. If it was manually configured,
 		// special testing may be occurring, and this may be indicated to just get a SAS-less URI.
 		// Alternatively, we may have already configured it here once before.
-		if c.Environment.AutoLoginMode == nil && c.Environment.ServicePrincipalAppID == nil && c.Environment.ServicePrincipalClientSecret == nil && c.Environment.AutoLoginTenantID == nil {
-			if GlobalConfig.StaticResources() {
-				c.Environment.AutoLoginMode = pointerTo("SPN")
-				oAuthInfo := GlobalConfig.E2EAuthConfig.StaticStgAcctInfo.StaticOAuth
-				a.AssertNow("At least NEW_E2E_STATIC_APPLICATION_ID and NEW_E2E_STATIC_CLIENT_SECRET must be specified to use OAuth.", Empty{true}, oAuthInfo.ApplicationID, oAuthInfo.ClientSecret)
+		if !c.Environment.ManualLogin {
 
-				c.Environment.ServicePrincipalAppID = &oAuthInfo.ApplicationID
-				c.Environment.ServicePrincipalClientSecret = &oAuthInfo.ClientSecret
-				c.Environment.AutoLoginTenantID = common.Iff(oAuthInfo.TenantID != "", &oAuthInfo.TenantID, nil)
-			} else {
-				// oauth should reliably work
-				oAuthInfo := GlobalConfig.E2EAuthConfig.SubscriptionLoginInfo
-				if oAuthInfo.Environment == AzurePipeline {
-					c.Environment.InheritEnvironment = true
-					c.Environment.AutoLoginTenantID = common.Iff(oAuthInfo.DynamicOAuth.Workload.TenantId != "", &oAuthInfo.DynamicOAuth.Workload.TenantId, nil)
-					c.Environment.AutoLoginMode = pointerTo(common.EAutoLoginType.AzCLI().String())
+			if c.Environment.AutoLoginMode == nil && c.Environment.ServicePrincipalAppID == nil && c.Environment.ServicePrincipalClientSecret == nil && c.Environment.AutoLoginTenantID == nil {
+				if GlobalConfig.StaticResources() {
+					c.Environment.AutoLoginMode = pointerTo("SPN")
+					oAuthInfo := GlobalConfig.E2EAuthConfig.StaticStgAcctInfo.StaticOAuth
+					a.AssertNow("At least NEW_E2E_STATIC_APPLICATION_ID and NEW_E2E_STATIC_CLIENT_SECRET must be specified to use OAuth.", Empty{true}, oAuthInfo.ApplicationID, oAuthInfo.ClientSecret)
+
+					c.Environment.ServicePrincipalAppID = &oAuthInfo.ApplicationID
+					c.Environment.ServicePrincipalClientSecret = &oAuthInfo.ClientSecret
+					c.Environment.AutoLoginTenantID = common.Iff(oAuthInfo.TenantID != "", &oAuthInfo.TenantID, nil)
 				} else {
-					c.Environment.AutoLoginMode = pointerTo(common.EAutoLoginType.SPN().String())
-					c.Environment.ServicePrincipalAppID = &oAuthInfo.DynamicOAuth.SPNSecret.ApplicationID
-					c.Environment.ServicePrincipalClientSecret = &oAuthInfo.DynamicOAuth.SPNSecret.ClientSecret
-					c.Environment.AutoLoginTenantID = common.Iff(oAuthInfo.DynamicOAuth.SPNSecret.TenantID != "", &oAuthInfo.DynamicOAuth.SPNSecret.TenantID, nil)
+					// oauth should reliably work
+					oAuthInfo := GlobalConfig.E2EAuthConfig.SubscriptionLoginInfo
+					if oAuthInfo.Environment == AzurePipeline {
+						c.Environment.InheritEnvironment = true
+						c.Environment.AutoLoginTenantID = common.Iff(oAuthInfo.DynamicOAuth.Workload.TenantId != "", &oAuthInfo.DynamicOAuth.Workload.TenantId, nil)
+						c.Environment.AutoLoginMode = pointerTo(common.EAutoLoginType.AzCLI().String())
+					} else {
+						c.Environment.AutoLoginMode = pointerTo(common.EAutoLoginType.SPN().String())
+						c.Environment.ServicePrincipalAppID = &oAuthInfo.DynamicOAuth.SPNSecret.ApplicationID
+						c.Environment.ServicePrincipalClientSecret = &oAuthInfo.DynamicOAuth.SPNSecret.ClientSecret
+						c.Environment.AutoLoginTenantID = common.Iff(oAuthInfo.DynamicOAuth.SPNSecret.TenantID != "", &oAuthInfo.DynamicOAuth.SPNSecret.TenantID, nil)
+					}
+				}
+			} else if c.Environment.AutoLoginMode != nil {
+				oAuthInfo := GlobalConfig.E2EAuthConfig.SubscriptionLoginInfo
+				if strings.ToLower(*c.Environment.AutoLoginMode) == common.EAutoLoginType.Workload().String() {
+					c.Environment.InheritEnvironment = true
+					// Get the value of the AZURE_FEDERATED_TOKEN environment variable
+					token := oAuthInfo.DynamicOAuth.Workload.FederatedToken
+					a.AssertNow("idToken must be specified to authenticate with workload identity", Empty{Invert: true}, token)
+					// Write the token to a temporary file
+					// Create a temporary file to store the token
+					file, err := os.CreateTemp("", "azure_federated_token.txt")
+					a.AssertNow("Error creating temporary file", IsNil{}, err)
+					defer file.Close()
+
+					// Write the token to the temporary file
+					_, err = file.WriteString(token)
+					a.AssertNow("Error writing to temporary file", IsNil{}, err)
+
+					// Set the AZURE_FEDERATED_TOKEN_FILE environment variable
+					c.Environment.AzureFederatedTokenFile = pointerTo(file.Name())
+					c.Environment.AzureTenantId = pointerTo(oAuthInfo.DynamicOAuth.Workload.TenantId)
+					c.Environment.AzureClientId = pointerTo(oAuthInfo.DynamicOAuth.Workload.ClientId)
 				}
 			}
-		} else if c.Environment.AutoLoginMode != nil {
-			oAuthInfo := GlobalConfig.E2EAuthConfig.SubscriptionLoginInfo
-			if strings.ToLower(*c.Environment.AutoLoginMode) == common.EAutoLoginType.Workload().String() {
-				c.Environment.InheritEnvironment = true
-				// Get the value of the AZURE_FEDERATED_TOKEN environment variable
-				token := oAuthInfo.DynamicOAuth.Workload.FederatedToken
-				a.AssertNow("idToken must be specified to authenticate with workload identity", Empty{Invert: true}, token)
-				// Write the token to a temporary file
-				// Create a temporary file to store the token
-				file, err := os.CreateTemp("", "azure_federated_token.txt")
-				a.AssertNow("Error creating temporary file", IsNil{}, err)
-				defer file.Close()
-
-				// Write the token to the temporary file
-				_, err = file.WriteString(token)
-				a.AssertNow("Error writing to temporary file", IsNil{}, err)
-
-				// Set the AZURE_FEDERATED_TOKEN_FILE environment variable
-				c.Environment.AzureFederatedTokenFile = pointerTo(file.Name())
-				c.Environment.AzureTenantId = pointerTo(oAuthInfo.DynamicOAuth.Workload.TenantId)
-				c.Environment.AzureClientId = pointerTo(oAuthInfo.DynamicOAuth.Workload.ClientId)
-			}
 		}
-
 		return target.URI(opts) // Generate like public
 	default:
 		a.Error("unsupported credential type")
@@ -339,8 +342,10 @@ func RunAzCopy(a ScenarioAsserter, commandSpec AzCopyCommand) (AzCopyStdout, *Az
 		0, command.ProcessState.ExitCode())
 
 	a.Cleanup(func(a ScenarioAsserter) {
-		UploadLogs(a, out, stderr, DerefOrZero(commandSpec.Environment.LogLocation))
-		_ = os.RemoveAll(DerefOrZero(commandSpec.Environment.LogLocation))
+		if !commandSpec.Environment.ManualLogin {
+			UploadLogs(a, out, stderr, DerefOrZero(commandSpec.Environment.LogLocation))
+			_ = os.RemoveAll(DerefOrZero(commandSpec.Environment.LogLocation))
+		}
 	})
 
 	return out, &AzCopyJobPlan{}
